Create code-of-conduct.md
This was using a sample of well known organizations that we will likely interact with. It appears many are using similar language, which is not surprising. Following my own experience and the advice of Metagov, this and other governance documents will attempt to define the least amount of rules and process until necessary to do so.

# Researched OSS communities codes of conduct
- CNCF: Good start on how to set broad rules to allow teams to self organize
- Gitcoin: very similar to CNCF with some good changes
- ENS: similar language to others. A bit too much detail that makes it harder to follow.
- Discord: a few service abuse references that were included
- Openstack (OpenInfra): very focused on conflicts of interest. We do not have enough commercial contribution to be worried about this yet.
- Giveth: very similar to the text already used